Move shared ELM response core
